свел в один контроллер
