--- lab-admin-ui/src/main/java/ru/optima/controller/ChiefController.java
@@ -1,17 +1,0 @@
-package ru.optima.controller;
-
-import org.springframework.security.access.annotation.Secured;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@Controller
-@Secured("ROLE_CHIEF")
-public class ChiefController {
-
-    @RequestMapping("/chief")
-    public String indexChiefPage(Model model) {
-        model.addAttribute("activePage", "None");
-        return "chief/index";
-    }
-}
